requirements.txt
flask==3.1.2
beautifulsoup4==4.14.2
requests==2.32.5
selenium==4.38.0
webdriver-manager==4.0.2
playwright==1.56.0
pillow==12.0.0
ImageHash==4.3.2
